requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.60.1
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
